tracing: Use process start time as unique identifier
Use the process start time hashed with the process id as the unique
identifier for a process. This ensures that if there are two independent
copies of the Perfetto SDK in the same process (e.g., one in the app and
another in a system framework), events emitted by each will be
consistently interleaved on common thread and process tracks.

Bug: b/178314761
Change-Id: Icccfd6813da6ed8bf9a6d8fbf9c41d964b65b5fb