Calculate threshold to treat allocation as sample.
Before this change, we would treat every allocation >= the sampling
interval as a sample. Had we done normal sampling on allocations ==
the sampling interval, only about ~63 % would contain at least one
sample.

Taking all of those did not bias the result, because we would sample
some of these 63 % multiple times. It did create a discontinuity in
the accuracy of the data, and would have us handle more samples than
we need to.

We now calculate the threshold where we would treat the allocation as a
sample 99% of the time.

Change-Id: Id748966dd4eba8f57bf7594e75057d36f6eb2c17