trace_processor: hackily allow empty callstacks in perf_sample table
We're adding counter-only samples, which will be emitted as either a
sample without a callsite_iid, or a callsite with zero frames (something
that the rest of the parsing doesn't assume is possible).

Relax the parser's assumption with a hacky fix (-1 is used elsewhere for
invalid IDs according to Lalit). The proper fix is likely to break the
inheritance of perf_sample table, and make the callsite id column
optional.

Bug: 180437026
Change-Id: I17004086306aafeb01f09f5a9364671638041801