UI: Disable Legacy UI after click and add spinner
Clicking on the "Legacy UI" button spawns a worker
which runs the conversion tool in the background.
There is little feedback for the user to tell that
something is happening.
This CL grays out the label, removes the click handler
and adds a spinner to give a visual indication that
something is happening.

Bug: 179429754
Change-Id: I9aa3a27dac21d34e43c2d82af0f8d7a1e28d6533