IPC: Allow socket to be TCP, for Windows
This CL adds a constants which defines whether a
AF_UNIX or AF_INET socket should be used in the IPC
layer. This CL alone doesn't add the TCP support,
just prepares for it. It also fixes the test_socket
to support both TCP and AF_UNIX..
Also fix a minor compiler issue with MSVC recently
regressed by the introduction of __has_attribute.

Bug: 174454879
Change-Id: Ic5ab727d33aac1d17db0b675b9a44cb132fefffa