UI: fix ellipsis in presence of wide code points
JS strings are UTF16 and substring can chop wider
codepoints in half, leading to invalid glyphs
(repro: jsfiddle.net/5o67r2sv).
This fix is correct (as in, it avoids codepoint
truncation and avoids text-overflow) but is
over-conservative, as it measures text assuming
1 wchar == width of 1 codepoint, which is not true.
As such might truncate eagerly text that has several
wide codepoints.
We can't possibly cope with broken emojis in tracing.

Bug: 181159104
Change-Id: Ife0c8f2afa347d03addc6772786c64b2e1155a74