UI: refactor TrackDecider logic
This is a pure no-op refactor of TrackDecider.
b/181255662 is still present with this CL.
The only purpose of this CL is to simplify the plumbing
of TrackDecider by using a stateful class, rather than
passing all the fields around to each function.
This is needed to put the codebase in a state that
makes it easier to fix b/181255662. The bug is NOT
fixed by this CL, will be done in a different CL
to split the refactoring from the behavioral change.

Bug: 181255662
Change-Id: I998e597c1ca31cd61f19e72df86893490a114843