UI: make flamegraph panel resizable again
The tooltip hover logic of flamegraph.ts was swallowing
the mousemove event. This made the details panel impossible
to resize down because as the pointer went over a flamegraph
slice, the mouse event was handled exclusively by the tooltip.
Removing the returns false makes the mousemove event bubble
up and handle the resize.

Change-Id: I8e9ea2269ddcb40f0c802eebcfe43f0118d60653