Update android_sysui_cuj to support new naming of slices
Choreographer#doFrame and DrawFrame slices will be renamed to include
the VSYNC id. Updating the scripts to handle that.
Not using the new id for matching slices to frames until we have more
devices updated.
Duplicated the test file to keep the tests with the old naming until we
are completly migrated

Bug: 161781469
Change-Id: I3ff7da9be33d914ed9d548064650102444b2b52a