traced_perf proto: move event definition to common/ + remove |tracepoint|
The successive changes will echo the descriptions of the counters
within the trace itself (as part of TracePacketDefaults), and it makes
sense to reuse the existing counter enum and Tracepoint message.

Root |config.tracepoint| is hard-deprecated as that support was added
not long before the config refactoring, so I'm not expecting any
dependents.

Bug: 180437026
Change-Id: Ie5c09b23a02c697d699bc45c74fc4076ad23a052