Update FrameTimeline actual slice colors
We want the color of the actual timeline slices to be based on the jank
type. Having this logic in the UI makes it restrictive when we want to
attribute colors to self/other component janks as there is very less
context on which component the slice belongs to. To address this, we
assign a jank tag to every slice at trace processor stage. The UI then
chooses the color to use for the specific jank tag.

Screenshot: https://screenshot.googleplex.com/ASHGhx6VM6thsh7

Bug: 181726082
Test: diff_test_trace_processor

Change-Id: I1e5d62aff521a6af2ccc8bd0cb2d07117d305f4c