TraceWriterImpl: Fix a DCHECK in drop_packets_ mode.
This DCHECK is only correct in SMB stall mode, but not in drop mode:
In drop mode, the patch list may be non-empty if the current chunk is
the garbage chunk (and thus invalid) and the last >=2 chunks contained
a fragment of the current packet and still need patching (and thus have
an entry in the patch list).

Note that only the DCHECK is invalid in this case (and will trigger a
spurious crash) -- the DCHECK's validity is not relied upon elsewhere.
Thus, production code already behaves correctly.

Change-Id: I7dd999a1850666eb4719383342cb8777e0bf5e3b