metrics: Change modified_rail_slices to a table
Many of the metrics operating on RAIL mode slices SPAN_JOIN the RAIL
slices with some other source. If the other input to SPAN_JOIN has N
partitions then it iterates over the RAIL slices operand N times. To
reduce the execution time for these queries, this changes
modified_rail_slices to a table so iterating over it repeatedly is not
so costly.

Bug: 185252076
Change-Id: Id325420b42c114eb7619992c7c1bb77b8b2b80bc