perfetto: add timerfd mechaism to snapshot periodic events in service
This CL implements timerfd based mechanism of snapshotting periodic
events to the tracing service.

The main usecase is to snapshot the clocks after long suspend
periods; because we expect large clock drift after suspend, it's crucial
that our clocks are resynchronized as soon after suspend as possible.
Using timerfd with the BOOTIME clock acts as a neat mechanism to do
this.

Bug: 185346474
Change-Id: I643fc32ee9651390f1d7e4703f2ec6691903cd10