ui: Immediately download trace from tools/record_android_trace
tools/record_android_trace works by recording a trace then (briefly)
hosting it on a local web server and opening a tab with:
https://ui.perfetto.dev/#!/?url=http://127.0.0.1:9001/foo.trace
As the UI loads, it fetches the trace from the local webserver. Once
this happens the webserver exits. The UI uses the 'URL' engine scheme
for these traces. If the user creates a permalink for a trace with this
scheme the trace is not uploaded - we assume that we can fetch it at
will from the URL. This is doubly wrong for traces sourced from
tools/record_android_trace. Firstly http://127.0.0.1 means something
different for each machine and secondly the webserver has exited and is
no longer running.

To avoid this problem we modify the ?url= handling code to immediately
download traces which come from 127.0.0.1 and thence use the File
engine scheme which will cause the trace to be uploaded if a
permalink is shared.

Bug: b/185235837
Change-Id: Icb41d95efba24d3f9c7aa8bc3d453882d53561a4