ui: Fix bug with grouping ion/dma tracks
Fix a bug where if there was at least one ion/dma
track but no suitable summary track we would crash.
The existing code attempted to handle this (by not adding the track
group) but did so incorrectly (since the tracks had beeen mutated to
know about the track group already). This moves the test of whether
we have sufficient tracks prior to the mutation.

Change-Id: I248f84d73bdc4e3c821f4f1e3d47bb5ae4c52b4e