Perfetto: Fix trace_processor HSC startup metrics for clock and gmail
Gmail and Clock both use animators as an indication that startup has
completed. Specifically, they look for the last frame of the longest
instance of a type of animation. As written, these metrics were looking
for the end of the latest initiated instance of an animation. Now they
look for the latest ending instance.

Bug: 187827061
Test: Ran trace_processor on a deskclock trace
Change-Id: Ia1bd2b6c08076c739478730e063371342560dc20