Extract long binder transactions during app startups.
The transactions are considered long if they are longer 100ms. We can probably refine this threshold later if we need to.

We output all long transactions individually. Based on a bit of ad-hoc analysis, the cardinality isn't generally very high, and if it is, we would really want to know about it.

For each transaction, we output the duration and the thread, which (sometimes) gives some idea on what kind of binder call it was. I haven't yet figured out a better way to trace binder transactions from traces, so will probably refine this in the future.

Change-Id: I31be44b40cabec56ad728cc3e3086dc82265da56