Client library: avoid re-creation of TLS on thread exit
The NonReentrantTaskRunner introduced by aosp/1642431
caused a subtle bug: the TLS object is re-created when
destroying the TracingTLS on thead exit. This is the call
stack that causes the issue:
- The at-thread-exit (the pthread_key_create dtor function)
  destroys the local TracingTLS instance.
- The TracingTLS dtor destroys the active TraceWriter instances
  for the various data sources.
- The TraceWriter dtor issues a PostTask on the IPC thread to
  flush the pending data an unregister its ID with the service.
- The PostTask now re-calls GetOrCreateThreadLocalObject() in
  two occasions:
  1. In the CallWithGuard in NonReentrantTaskRunner (This alone
     would be easy to fix with a GetWithoutCreateTls()).
  2. In Chromium, the TaskRunner has TRACE_EVENTs that will try
     to emit trace points (which is why CallWithGuard has been
     introduced in the first place).

This CL fixes the issue by introducing a sentinel when destroying
the TLS object and returning a global fake TLS instance to the
GetOrCreateThreadLocalObject() made during the dtor when the
sentinel is active.
This is safe only because all the code paths that re-enter tracing
have an early out when |is_in_trace_point| == true.

Bug: 180295966
Change-Id: I78dbdab375ef46c32853e20ecac280e8b78e9bdc