fixit: metrics: add support for strongly typed enum fields in metrics
This CL adds the ability to specify enum values by using a string of
their name rather than the integer value. This expands the support added
in aosp/1661905.

We also strengthen type-checking of enum values to make sure they are in
bounds and migrate the 'trace_stats' metric to using enum strings rather
than values.

Change-Id: I7a6ff12b178373c0ddb078ec329ae1050971d621
Bug: 184265484