UI: Fix ADB/USB connection issues on Mac/Windows
Resetting the interface (with the intent of stealing it from
ADB host) causes problems on Mac in more recent versions of
Chrome (probably is also the cause of similar problems
also on Windows). Remove the reset().
This means that if a user has an ADB session ongoing they
need to first stop it as instructed (our message improved in
this regard).

Bug: 189971281
Test: manual on Mac and windows
Change-Id: I4fbbea9669d9b2af3fb48c7a1876a9c6a774de48