tools: add skeleton of proto merger tool
This tool allows for "syncing" a .proto file by reading the contents of
a "source of truth" proto and merging any changes in an intellgient
manner.

This CL introduces the initial logic for reading .proto files from disk
and converting them into a struct amenable for easy merging. Follow up
CLs will implement the actual merging algorithm + seralization back to a
.proto file.

Change-Id: Icb189c238597a32f25cbd5d423e7c3a2936b1d51
Bug: 188053828