perfetto_cmd: move go/trace-uuid print after incidentd checks
Today the go/trace-uuid is printed before calling
SaveTraceIntoDropboxAndIncidentOrCrash().
This leads to a particularly unfortunate scenario where
invoking perfetto --save-for-bugreport (which happens on every
bugreport) prints a link that will 100% NOT be valid, because
that trace will later fail with a
"Skipping write to incident. Empty trace".
Moving the uuid printing after the empty check.

Bug: 185096394
Change-Id: I065a88e818bf45f3805e3b065462e161e5900fcb