ui: Move trace conversion to dedicated worker
Trace conversion currently takes place synchronously on the controller
worker. We plan to remove the controller worker and move that code to
the main thread however the trace conversion can not be run synchronously
on the main thread so we must first move the trace conversion to a
dedicated worker.

Change-Id: Ic53fee568bf107c4b6bc4f1bf956376702c04648