perf proto: allow for raw event timebases, and named counters
The raw event specifies the raw values for {type, config, config1,
config2} fields of the perf_event_attr struct, which is used in the
perf_event_open syscall. Typically an event is specified just by |type|
and |config|. This should help make the profiler more flexible for local
use-cases. The ergnomics aren't great, as it'll require the user to
manually encode the counters in question, but it should allow for
sampling almost every perf counter. Proper symbolic names and
discovering of dynamic PMUs would be nice, but there are more important
issues to solve in the near term.

The other addition is allowing the timebase to be given an arbitrary
name via the config. This does not affect the profiling, but helps
identify the relevant data stream within the trace at parse time.

Bug: 192643545
Change-Id: Ie8103281aac3443832036b50d5642e0ca729f4cf