tp: parse raw and explicitly named perf counters
The only changes are in naming of counter tracks. If an explicit name is
given by the config, it is used for the track's name. Otherwise the
parser chooses an appropriate name.

For raw counters, I don't have a particular preference for the
automatic naming. I would hope that the default name won't be relied
upon too much, given that we're adding an option for explicit naming.

Bug: 192643545
Change-Id: Iff11430a548af2165a657b9b460a5092008868a1