ftrace: turn binary ftrace parsing errors into crashes
traced_probes should always know how to parse the kernel-written binary
ftrace format unless there is a bug, or us not keeping up with a
significant kernel change. We must detect these issues as soon as
possible, so this patch adds a CHECK that crashes the process if we fail
to parse an ftrace page.

An alternative approach would've been to add a new bool to the ftrace
stats proto, which could be written at the end-of-trace stat packet, and
then turned into a data loss error in the trace processor. I've
considered both approaches, and concluded that it's much more important
to have as loud of a signal as possible (crash), to avoid these issues
being unnoticed due to relevant traces not getting inspected.

This does pose a possibility of crashlooping while continuously
attempting the same ftrace tracing session across restarts, as long as
an affected tracing config is online.

For the record, if we decide to go back to the ftrace stats approach:
the best way I've found is to add a separate bool to FtraceDataSource,
and have it be set by CpuReader::ReadAndProcessBatch (instead of the
current CHECK). The FtraceMetadata was not an appropriate struct for
this, as it does not have any state that is designed to stay valid until
the end of the session.

Bug: 191160035
Change-Id: I2b93837ecdc69082e648a4b3c263832530fd4bd1