Add tools/roll-prebuilts
Adds a script which takes the LUCI prebuilts for the various
platforms and updates the tools/{trace_processor, traceconv}
scripts, updating the URL and SHA-256 pinning of each of them.
This guarantees that the Perfetto GIT repository remains the
source of trust for resource integrity.

I didn't run the script yet so the current prebuilt scripts
in the repo are still stale.

Change-Id: Ib488cdfcecc465485d853e3c2258cea48248e7d9
Bug: 177349647