Modified buffer length in breakpad_parser.cc.
It was assumed that function names that the BreakpadParser read would
not be longer that 1000 characters. This assumption was incorrect and
could cause segmentation faults. This change uses dynamic character a
dynamic character array in breakpad_parser.cc and a std::string to
append to in breakpad_symbolizer.cc, instead of constant character
arrays.

Bug: b/191080142
Change-Id: I7b458aecc7f5f791392e5b52623431aacb6e6cd3