UI: allow duplicate column names in result sets
In some rare case a result set can have two columns with
the same name (e.g., joins or simply `select 1 as x, 2 as x`).
In this case, the QueryResult iterator has an unexpected behavior,
that is, creates a {x: 2} object propagating only the last column.
Keep the old (pre-QueryResult) behavior that renamed columns if
they clash. It's not great but there isn't much we can do at the
UI level.

Bug: 194891824

Test: ui/run-unittests
Test: manual, open query page type `select 1 as x, 2 as x`.
Change-Id: I67f33901161c1bcb69d511e376f86f584491aa23