metrics: Switch shell prototext output to ComputeMetricText
In prepration for removing libprotobuf descriptorpool usage from trace
processor shell in order to support text output of extension metrics,
we're switching the prototext output to ComputeMetricText which uses the
internal DescriptorPool instead of the libprotobuf DescriptorPool.

There might be some changes in prototext output; for example, fields
may be output in a different order.

Bug: 182165266
Change-Id: I544047092e3e1717c519d8b82f3f976cd9f12fcc