ui: Remove route from state (take 2)
This follows the design of go/perfetto-ui-router-refactoring.
Since the controllers now run in the main thread, the majority of
usecases for keeping the route in the state go away.
Therefore, we can eliminate state.route and Actions.navigate.
We can also eliminate logic which keeps state.route and window.location
in sync.

Instead, we will now have a callback that listens to URL changes.
This callback will handle 8 cases of URL transitions, documented
in the code below.

Test: manual, see last section of go/perfetto-ui-router-refactoring
Change-Id: I9d5c1c3c0479dd783a41402e5a8ad47a5e81bd53