service: Fix reporting of producer uid in --query
An oversight in the code that populates the TracingServiceState
(the proto returned by --query and --query-raw) caused the producer
UID to be always 0.
The bug was about reading the UID from the proto object that we
were intending to populate, rather than from the ProducerEndpoint
object that stores the uid.

Test: TracingServiceImplTest.QueryServiceState
Bug: 197190513
Change-Id: Ia5a9f6d3b25f1074a325c8a02e6987846f5fdf49