Change release version format to include SHA1
The current version number format is:

v{Major version}.{Minor version}.{# of new commits}

which ends up looking like this: v18.0.12

Unfortunately this is ambiguous with respect to branches since v18.0.12
could mean 12 commits on main since v18 or 12 cherry-picks on a branch
based on v18. To resolve this we're modifying the version format to
include the SHA1 of the commit at which it was built. The new format is:

v{Major version}.{Minor version}-{first 9 chars of the commit SHA1}

which ends up looking like this: v18.0-8420673b4

Bug: 198293659
Change-Id: I6fa6de2e14db83254b12cf2d3d81466e5de2c609