[Pivot Table] Improve UI design of pivot tables
Improved the UI flow of pivot tables, added a pop up window to edit
pivot and aggregation selection, drag and drop to reorder columns in the
pop up window and table, and an option to sort aggregations from the
table.
Next Step: Modify query generation and response to support expandable
tables.

Bug: 149193812
Change-Id: Ifeef60cfbe61effa8534b321cc6b02e841867a45