tracing: remove 0-dchecks on patches on the service
Until this CL the codebase was requiring that if the tracing
service was built in debug mode (i.e. without -DNDEBUG) then
all the producers were too.
This caused the tracing protocol ABI to depend on the -DNDEBUG
level and later turned out to be a source of problems for
users of the SDK.
Since the beginning of the project, this DCHECK has never found
any issue, hence we are removing it to remove the ABI dependency
on -NDEBUG.

Bug: 197340286
Change-Id: I8d6cf1374acd245407e660fd8e761d01039dc0fc