Avoid using fts_open and fts_read
b39ae485a46b("Link against libfts when compiling against musl") had to
introduce an additional dependency on libfts when building with musl as
libc.

Turns out that base::ListFilesRecursive() can be used instead. This has
two benefits:
* Doesn't require libfts when building with musl.
* Reduces platform dependend duplicated code.

perfetto::profiling::WalkDirectories() used a callback, while
base::ListFilesRecursive() returns a vector, so there might be a
performance difference with very large trees. If this turns out to be a
problem, base::ListFilesRecursive() can be changed to use a callback (or
return an iterator instead).

Bug: 200136875
Change-Id: I420e7b0ad7fe10666da9c39baf1ca046649588fd