Add load benchmark for kallsyms
There already is a benchmark for lookups. This instead measures the time
it takes to parse and build the data structure (it should only happen
once per trace).

The benchmark is not very representative because reading from a real
file is very different than reading from /proc/kallsyms (generated on
the fly by the kernel), but it can still be useful to compare different
approaches.

Bug: 180093708
Change-Id: I5747d3a483bdaa2c192d9f732a19df9d87b6e37b