Avoid windows.h in header files
proc_utils.h and thread_utils.h both used to include windows.h and
processthreadsapi.h in order to get the prototype for a single function.
This causes compilation overhead and significant namespace pollution and
can be easily avoided. This change puts the function prototypes inline,
compatible with windows.h.

Compiling with this change found one source file in Chromium that was
using windows.h functions without including windows.h so a fix to that
(see crrev.com/c/3183604) must land first (and has).

Fixing this lowers the number of translation units that include
windows.h in a build of Chrome by 244, or almost 9% of what remains.

Bug: 796644
Change-Id: Ib75f2ef9f25996bae8795b584abddf14cb6befd3