Fix ftrace imports for hosts w/ > 128 CPUs
We're trying to evaluate hardware that has more than 128 CPUs, and this change
fixes the hard-coded CPU limit for ftrace imports.

I haven't done any benchmarks but this change should make
SchedEventTracker::FlushPendingEvents() slightly faster on hosts with < 128
CPUs (which I suspect is most hosts), since this the for loop in this method
will now only have as many loop iterations as there are actual CPUs.

Change-Id: I63a01ad5040d1abf96665f86bfee7f6bffe8bd70