base: Add LogRingBuffer to keep around last logs
This class will be used in loggging.cc by upcoming CLs
to keep the last PERFETTO_LOG messages around and output
them in the case of a managed crash.
This CL adds only the basic ring buffer and a test and
does not alter the behavior of logging.cc, which will
happen in follow ups.

Bug: 202283508
Change-Id: I0661ef1dd762c7b408776b3034f40815052d0964