Expand "Support" by default
Compared to Catapult, Perfetto UI lacks immediately visible "?" button
showing basic controls. The corresponding menu item exists and available
when clicking "Support" group, but user has to experiment with the UI in
order to be able to find it, which has been a problem in practice (the
problem has been brought up by a user in response to request for
feedback).

This CL makes the "Support" section expanded by default and swaps it
with "Sample queries"; without the swap "Sample queries" is collapsed
in-between expanded section which makes the menu look a bit weird.


Bug: 204404247
Change-Id: I8aaad88ed13a504f67cda9226b83d78133fa04a4