Request categories at most once
Controller's run() method is called twice in rapid succession, which
leads to concurrency issues in the extension that ultimately makes
the first recording attempt to fail. This CL makes the UI request
categories list at most once, avoiding the error.

While this is not a proper fix to extension's concurrency issues, it
should make the recording of Chrome traces from Perfetto UI work for
the time being.

Bug: 199997583
Change-Id: I3a9cd67d05add6d0ef490468ee61a6f00dd065e7