Tracing protocol: introduce UpdateDataSource
This CL is a follow up to aosp/1818477 and moves the logic
for updating a data source descriptor into a dedicated
RPC endpoint.
The previous CL was not backwards-compatible. When a newer
client targets an older service (which doesn't have the
update semantics), the update requests would have looked
like new registrations.
This CL introduces a dedicated method, which will gracefully
fail without side effects on older versions of the service.

Bug: 180347809
Change-Id: I0a68a4a98f01cd5bd7fea9b45fe0a92105c547ac