Fix flaky integration test.
The test was accessing pages by url and then
clicking 'Back' and 'Forward'.

However, the pages were not recorded in stack so 'Back' and 'Forward'
had no effect.

Also, there were extraneous 'Back' and 'Forward'.

Added an id for each button, so it can be more easily selected by
pupeteer.


Change-Id: Ie26473beaf48196e6ae17460d2a05037c1352e65