Do not render flow arrow heads for sufficiently short arrows
Profiling shows that rendering arrow heads takes a significant amount
of time (close to 20% in recorded profiles!) when a lot of flow
arrows are present on a screen and the viewport is significantly
zoomed out. At that zoom level, individual arrows are not easily
distinguishable anyway, therefore this CL stops rendering arrow heads
when difference in X coordinate is sufficiently small.

Bug: 205265022
Change-Id: I86acb528888756f70a33cfd4cb82475c8f89d655