[fuchsia] Fix GetoptCompatTest/1.LongOptions flakes.
GetoptCompatTest were expecting that getopt and getopt_long would set
optarg to null unconditionally if the option under test did not have an
argument. It seems that getopt_long on Fuchsia only sets optarg to null
if the relevant option in optstring is followed by a colon or if the
getopt_long option specifies that an argument is optional or
required. The manpage for getopt_long(3)
(https://linux.die.net/man/3/getopt_long) doesn't clearly state that
optarg is set to null in cases where the option is not followed by a
colon. Since it doesn't make sense to check optarg for options that
don't accept or require an arg, let's stop testing for optarg == nulptr.
This fixes GetoptCompatTest/1.LongOptions, which failed on all Fuchsia
runs and then succeeded on the first retry because optarg still pointed
to a memory location that was relevant in a previous test.

Bug: None
Change-Id: Icba0e99baee11cb0e30dfbf97297fbcec64c6805