Populate waker_utid arg for sched_waking and sched_wakeup
For sched_wakeing and sched_wakeup, trace processor is populating the
`instant` table. The `ref` column is used to store the utid that's being
woken up.

The running utid when the event happens (i.e. the waker) was not stored
in the instant table. Indeed, the ui was still using the raw table to
get this info.

With this commit, when instant.name is "sched_waking" or "sched_wakeup",
the "waker_utid" argument reports the running utid when the event
happened.

Bug: 206015754
Change-Id: I80863a68f9ab2b78158d85afa04229975b78cdd4