perfetto/rss_stat_tracker: Use rss_stat_throttled if enabled
Update the ftrace parser to use rss_stat_throttled event if enabled.
It doesn't make sense to specify both rss_stat and rss_stat_throttled
in the same config; if both are enabled then the throttling is
ineffective. So we can reuse the rss_stat tracker to parse both rss_stat
and rss_stat_throttled events.

Bug: 145972256
Test: - tools/install-build-deps --ui
      - ui/build -ws
      - Upload trace on local server (http://127.0.0.1:10000)
      - Observe mem.rss.file, mem.rss.anon, mem.rss_shmem metrics

Change-Id: Id56481f3eade7f2908d3d3bce4f6765ca94e2adf