proto: traced_perf: allow explicitly setting a clockid for perf timestamps
Introducing a separate clock enum besides BuiltinClock since perf events
support only a specific subset of the clocks.

The backing intention is to go back to using CLOCK_BOOTTIME as much as
possible, to avoid clock skew relative to other trace packets around
suspend, since the resynchronisation of monotonic and boot clocks is not
immediate (done periodically in userspace).

Change-Id: I36e36b1ccab25f0c43837909d77bc14323bcec55