Make the undo/redo work for programmatic changes
Turns out one does not just modify textarea's value attribute: this
makes the element lose its edit history, breaking undo and redo
functionality. In order to keep the edit history, document.execCommand
can be used instead.

This CL also fixes a couple of minor issues with Tab handling for
multiline selection: if only the last character was selected on the
first line of selection, the first line was not indented. Also, the
selection is restored starting from the beginning of modified line
instead of trying to keep the beginning at the same character (which did
not work and is complicated to implement properly): while being less
accurate, it still allows indenting several times by doing selection
once and pressing Tab several times.

Change-Id: I96d9d4e2f104cfdd097ec11386f77cf4c3b32101