Allow grouping of derived event tracks
This is a feature that allows derived event tracks to be placed under
a track group, as an attempt to better support metrics that yield
multiple tracks (e.g. chrome_dropped_frames).

This CL adds support for an optional |group_name| column in the
[metric_name]_event tables. All tracks with the same |group_name| will
be placed under the group by that name. The first track inserted in
the [metric_name]_event table will be used as the summary track for
the group.

Bug: 204226400
Change-Id: I3b3dd178e036d7497c5d6dfbb0476a502399f981